requirements.txt
fastapi
gunicorn
huggingface-hub
numpy
pydantic
python-dotenv
uvicorn
transformers
peft
torch